Move structs about expectations into controller/expectations package
