+++ controller/expectations/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "expectations",
+    srcs = ["expectations.go"],
+    importpath = "github.com/kkohtaka/kubernetesimal/controller/expectations",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@io_k8s_apimachinery//pkg/util/sets",
+        "@io_k8s_client_go//tools/cache",
+        "@io_k8s_klog_v2//:klog",
+        "@io_k8s_utils//clock",
+    ],
+)
