+++ controllers/BUILD.bazel
@@ -3,14 +3,12 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "controllers",
     srcs = [
-        "errors.go",
         "etcd.go",
         "etcd_controller.go",
         "etcd_prober.go",
         "etcdnode.go",
         "etcdnode_controller.go",
         "etcdnode_prober.go",
-        "expectations.go",
         "finalizer.go",
         "pki.go",
         "ssh.go",
@@ -25,6 +23,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//api/v1alpha1",
+        "//controller/errors",
+        "//controller/expectations",
         "//k8s/endpointslice",
         "//k8s/etcdnode",
         "//k8s/object",
@@ -43,10 +43,7 @@ go_library(
         "@io_k8s_apimachinery//pkg/runtime",
         "@io_k8s_apimachinery//pkg/types",
         "@io_k8s_apimachinery//pkg/util/intstr",
-        "@io_k8s_apimachinery//pkg/util/sets",
-        "@io_k8s_client_go//tools/cache",
         "@io_k8s_client_go//util/workqueue",
-        "@io_k8s_klog_v2//:klog",
         "@io_k8s_sigs_controller_runtime//:controller-runtime",
         "@io_k8s_sigs_controller_runtime//pkg/client",
         "@io_k8s_sigs_controller_runtime//pkg/controller/controllerutil",
@@ -54,7 +51,6 @@ go_library(
         "@io_k8s_sigs_controller_runtime//pkg/handler",
         "@io_k8s_sigs_controller_runtime//pkg/log",
         "@io_k8s_sigs_controller_runtime//pkg/source",
-        "@io_k8s_utils//clock",
         "@io_k8s_utils//pointer",
         "@io_kubevirt_api//core/v1:core",
         "@io_opentelemetry_go_otel_trace//:trace",
@@ -63,16 +59,12 @@ go_library(
 
 go_test(
     name = "controllers_test",
-    srcs = [
-        "errors_test.go",
-        "suite_test.go",
-    ],
+    srcs = ["suite_test.go"],
     embed = [":controllers"],
     deps = [
         "//api/v1alpha1",
         "@com_github_onsi_ginkgo//:ginkgo",
         "@com_github_onsi_gomega//:gomega",
-        "@com_github_stretchr_testify//assert",
         "@io_k8s_client_go//kubernetes/scheme",
         "@io_k8s_sigs_controller_runtime//pkg/client",
         "@io_k8s_sigs_controller_runtime//pkg/envtest",
